requirements.txt
streamlit
pdfplumber
langchain
openai
faiss-cpu
pdfplumber
langchain-community
tiktoken
sentence_transformers